add board, optimization ControlBoards and change name ControlBoard => ControlBoards
