--- src/components/Sidebar/ControlBoard/index.js
@@ -1,1 +1,0 @@
-export { default } from './ControlBoard';
