+++ src/components/Sidebar/ControlBoards/ControlBoards.styled.js
@@ -71,82 +71,3 @@ export const UlStyled = styled.ul`
 export const TrackVertical = styled.div`
   background-color: transparent;
 `;
-
-export const LiStyled = styled.li`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  width: 100%;
-  height: 70px;
-  transition: background-color ${baseTransition};
-  cursor: pointer;
-
-  &.active {
-    background-color: ${props => props.theme.colors.activeBoardBackground};
-  }
-`;
-
-export const DivNameStyled = styled.div`
-  display: flex;
-  align-items: center;
-  column-gap: 4px;
-  margin-left: 14px;
-  margin-right: 14px;
-
-  color: ${props => props.theme.colors.boardTitleTextIcon};
-
-  &.active {
-    color: ${props => props.theme.colors.sideBarMainText};
-  }
-  @media screen and (${devices.tablet}) {
-    column-gap: 8px;
-    margin-left: 24px;
-  }
-`;
-
-export const TextStyled = styled.p`
-  font-family: Poppins;
-  letter-spacing: -0.02em;
-  width: 120px;
-`;
-
-export const DivIconStyled = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-  column-gap: 8px;
-  padding-right: 18px;
-  height: 100%;
-
-  &::after {
-    content: '';
-    position: absolute;
-    right: 0;
-
-    width: 4px;
-    height: 96%;
-    border-radius: 8px 0 0 4px;
-    background-color: ${props => props.theme.colors.btnLogOut};
-  }
-`;
-
-export const ButtonEdit = styled.button`
-  color: ${props => props.theme.colors.boardTitleTextIcon};
-  transition: color ${baseTransition};
-
-  &:hover,
-  &:focus {
-    color: ${props => props.theme.colors.sideBarMainText};
-  }
-`;
-
-export const ButtonDelete = styled.button`
-  color: ${props => props.theme.colors.boardTitleTextIcon};
-  transition: color ${baseTransition};
-
-  &:hover,
-  &:focus {
-    color: ${props => props.theme.colors.sideBarMainText};
-  }
-`;
