+++ src/hooks/useModal.js
@@ -11,15 +11,3 @@ export const useModal = (initialState = false) => {
   const closeModal = () => setIsModalOpen(false);
   return { isModalOpen, openModal, closeModal };
 };
-
-export const useEditModal = (initialState = false) => {
-  const [isModalEditOpen, setIsModalEditOpen] = useState(initialState);
-  useEffect(() => {
-    if (isModalEditOpen) {
-      document.body.style.overflow = 'hidden';
-    }
-  }, [isModalEditOpen]);
-  const openEditModal = () => setIsModalEditOpen(true);
-  const closeEditModal = () => setIsModalEditOpen(false);
-  return { isModalEditOpen, openEditModal, closeEditModal };
-};
